+++ ShiftCinema/package.json
@@ -18,6 +18,8 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.29.0",
+    "@svgr/core": "^8.1.0",
+    "@svgr/webpack": "^8.1.0",
     "@types/react": "^19.1.8",
     "@types/react-dom": "^19.1.6",
     "@vitejs/plugin-react": "^4.5.2",
@@ -27,6 +29,7 @@
     "globals": "^16.2.0",
     "typescript": "~5.8.3",
     "typescript-eslint": "^8.34.1",
-    "vite": "^7.0.0"
+    "vite": "^7.0.0",
+    "vite-plugin-svgr": "^4.3.0"
   }
 }
