+++ ShiftCinema/src/App.tsx
@@ -1,4 +1,3 @@
-import './App.css'
 import {BrowserRouter, Navigate, Route, Routes} from "react-router";
 import HomePage from "./pages/HomePage.tsx";
 
@@ -6,7 +5,7 @@ function App() {
     return (
         <BrowserRouter>
             <Routes>
-                <Route path={"/"} element={<Navigate to="/cinema/today" replace/>}/>
+                <Route path={"/"} element={<Navigate to={"/cinema/today"} replace/>}/>
                 <Route path={"/cinema/today"} element={<HomePage />} />
             </Routes>
         </BrowserRouter>
