+++ ShiftCinema/src/components/FilmCard.tsx
@@ -1,22 +1,24 @@
-import type {Film} from "../api/filmsApi.ts";
+
 import * as React from "react";
 import {useNavigate} from "react-router";
 import "../styles/class-card-style.css"
+import type {Film} from "../data/filmResponse.ts";
 
 interface FilmProps {
     film: Film
 }
 
+const ratingMap : {[key: string] : string} = {
+    "G": "0+",
+    "PG": "6+",
+    "PG13": "12+",
+    "R": "16+",
+    "NC17": "18+",
+}
+
 const FilmCard: React.FC<FilmProps> = ({film}) => {
     const navigate = useNavigate()
     const pathImg = "https://shift-intensive.ru/api" + film.img
-    const ratingMap : {[key: string] : string} = {
-        "G": "0+",
-        "PG": "6+",
-        "PG13": "12+",
-        "R": "16+",
-        "NC17": "18+",
-    }
 
     return (
         <div className={"class-card"}>
@@ -29,7 +31,7 @@ const FilmCard: React.FC<FilmProps> = ({film}) => {
             </div>
 
 
-            <h3 className={"film-title"}>{film.name}({ratingMap[film.ageRating]})</h3>
+            <h3 className={"film-title"}>{film.name} ({ratingMap[film.ageRating]})</h3>
             <p className={"film-type"}>Фильм</p>
             <button className={"detail-button"} onClick={() => navigate(`/cinema/${film.name}`)}>Подробнее</button>
         </div>
