+++ ShiftCinema/src/components/Header.tsx
@@ -1,4 +1,9 @@
 import {useNavigate} from "react-router";
+import Logo from "../assets/logo.svg?react"
+import UserIcon from "../assets/user_icon.svg?react"
+import TicketIcon from "../assets/ticket_icon.svg?react"
+import ExitIcon from "../assets/exit_icon.svg?react"
+import SwitchMode from "../assets/switch_mode.svg?react"
 import "../styles/header-style.css"
 
 const Header = () => {
@@ -8,63 +13,30 @@ const Header = () => {
         <header className={"header"}>
             <div className={"header-container"}>
                 <span onClick={() => navigate("/")} className={"logo-span"}>
+
                     <div>
                         <p>Шифт</p>
                         <p>Cinema</p>
                     </div>
-                    <svg width="51" height="32" viewBox="0 0 51 32" fill="none" xmlns="http://www.w3.org/2000/svg">
-                        <path
-                            d="M18.5629 16.9104H15.9433C14.8276 16.9104 13.922 17.816 13.922 18.9317C13.922 20.0475 14.8276 20.953 15.9433 20.953H19.1247C19.7938 20.953 20.3768 21.394 20.6169 22.0184C21.2478 23.6596 22.1491 25.1671 23.2658 26.4859C23.8573 27.184 23.3695 28.2539 22.4546 28.2539H19.7447C18.7772 28.2539 17.943 28.9977 17.8972 29.9652C17.8487 30.9974 18.6721 31.8504 19.6935 31.8504H32.9881C33.6942 31.9488 34.4138 32 35.1469 32C43.8088 32 50.8199 24.9215 50.7081 16.2313C50.5989 7.82947 43.74 0.977319 35.3382 0.874908C34.1605 0.860085 33.0124 0.977319 31.9074 1.21179C29.2663 1.77236 26.5699 2.03108 23.8708 2.03108H10.9858C9.89698 2.03108 8.95911 2.86789 8.9079 3.95669C8.85265 5.11826 9.7784 6.0777 10.9278 6.0777H19.2084C20.9184 6.0777 22.0341 7.88203 21.2606 9.40608C21.2512 9.4236 21.2431 9.44111 21.2337 9.45863C20.8456 10.2335 20.0506 10.7213 19.1828 10.7213H8.58719C7.49839 10.7213 6.56051 11.5581 6.50931 12.6469C6.45406 13.8084 7.37981 14.7679 8.52925 14.7679H18.5616C19.1693 14.7679 19.6544 15.2718 19.6315 15.8796C19.6099 16.455 19.1369 16.9091 18.5616 16.9091L18.5629 16.9104ZM30.7822 7.10586C30.9224 4.92961 32.6903 3.16435 34.8666 3.02825C37.5064 2.86385 39.6799 5.03875 39.5128 7.67855C39.3754 9.85615 37.6088 11.6228 35.4298 11.7602C32.786 11.9273 30.6098 9.747 30.7809 7.10451L30.7822 7.10586ZM22.0301 14.0752C21.867 11.4354 24.0419 9.26324 26.6817 9.43033C28.8607 9.56778 30.6273 11.3357 30.7634 13.5147C30.9291 16.1585 28.7475 18.3334 26.105 18.1623C23.9287 18.0208 22.1648 16.2528 22.0287 14.0766L22.0301 14.0752ZM33.8789 24.1251C33.7414 26.3027 31.9748 28.0693 29.7972 28.2067C27.1534 28.3738 24.9744 26.1935 25.1456 23.551C25.287 21.3734 27.055 19.6095 29.2326 19.4734C31.8724 19.309 34.0446 21.4853 33.8789 24.1251ZM35.3989 17.8699C34.4057 18.0356 33.5568 17.1907 33.7171 16.1976C33.8142 15.5966 34.2966 15.1101 34.8976 15.0091C35.8907 14.8433 36.7423 15.6869 36.582 16.6814C36.485 17.2837 36.0012 17.7702 35.3989 17.8712V17.8699ZM45.1522 24.1224C45.0161 26.3 43.2509 28.0679 41.0733 28.2067C38.4294 28.3765 36.2505 26.1962 36.4203 23.5537C36.5591 21.3775 38.3243 19.6109 40.5006 19.4734C43.139 19.3063 45.3166 21.4826 45.1522 24.1224ZM43.9004 18.1717C41.4829 18.1717 39.5236 16.2124 39.5236 13.7977C39.5236 11.3829 41.6541 9.25515 44.213 9.43303C46.3717 9.5826 48.1195 11.333 48.2664 13.4918C48.4402 16.048 46.4176 18.1731 43.9017 18.1731L43.9004 18.1717Z"
-                            fill="#712D9C"/>
-                        <rect x="10.7734" y="28.3206" width="5.42906" height="3.61937" rx="1.80969" fill="#712D9C"/>
-                        <rect x="0.06604" y="2.08041" width="6.78633" height="3.92099" rx="1.96049" fill="#712D9C"/>
-                    </svg>
+                    <Logo />
                 </span>
                 <div className={"desktop-navigation"}>
                     <span onClick={() => navigate(`/account/${id}`)} className={"desktop-navigation-span"}>
-                        <svg width="24" height="24" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
-                            <path fillRule="evenodd" clipRule="evenodd"
-                                  d="M16 8C16 10.2091 14.2091 12 12 12C9.79086 12 8 10.2091 8 8C8 5.79086 9.79086 4 12 4C14.2091 4 16 5.79086 16 8ZM14.5 8C14.5 9.38071 13.3807 10.5 12 10.5C10.6193 10.5 9.5 9.38071 9.5 8C9.5 6.61929 10.6193 5.5 12 5.5C13.3807 5.5 14.5 6.61929 14.5 8Z"
-                                  fill="#141C24"/>
-                            <path fillRule="evenodd" clipRule="evenodd"
-                                  d="M4 17.772V20H20V17.772C20 16.1046 18.9657 14.6121 17.4045 14.0267L17.2669 13.9751C13.8711 12.7017 10.1289 12.7017 6.73315 13.9751L6.59551 14.0267C5.03429 14.6121 4 16.1046 4 17.772ZM7.25984 15.3796C10.316 14.2335 13.684 14.2335 16.7402 15.3796L16.8778 15.4312C17.8536 15.7971 18.5 16.7299 18.5 17.772V18.5H5.5V17.772C5.5 16.7299 6.14643 15.7971 7.12219 15.4312L7.25984 15.3796Z"
-                                  fill="#141C24"/>
-                        </svg>
+                        <UserIcon />
                         <p>Профиль</p>
                     </span>
                     <span className={"desktop-navigation-span"}>
-                        <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-                            <path
-                                d="M12 17C12.5523 17 13 16.5523 13 16C13 15.4477 12.5523 15 12 15C11.4477 15 11 15.4477 11 16C11 16.5523 11.4477 17 12 17Z"
-                                fill="#141C24"/>
-                            <path
-                                d="M13 12C13 12.5523 12.5523 13 12 13C11.4477 13 11 12.5523 11 12C11 11.4477 11.4477 11 12 11C12.5523 11 13 11.4477 13 12Z"
-                                fill="#141C24"/>
-                            <path d="M12 9C12.5523 9 13 8.55228 13 8C13 7.44772 12.5523 7 12 7C11.4477 7 11 7.44772 11 8C11 8.55228 11.4477 9 12 9Z"
-                                  fill="#141C24"/>
-                            <path fillRule="evenodd" clipRule="evenodd"
-                                  d="M2 14V18C2 19.1046 2.89543 20 4 20H20C21.1046 20 22 19.1046 22 18V14C20.8954 14 20 13.1046 20 12C20 10.8954 20.8954 10 22 10V6C22 4.89542 21.1046 4 20 4H4C2.89543 4 2 4.89545 2 6V10C3.10457 10 4 10.8954 4 12C4 13.1046 3.10457 14 2 14ZM20.5 18C20.5 18.2761 20.2761 18.5 20 18.5H4C3.72385 18.5 3.5 18.2761 3.5 18V15.1632C4.68246 14.6015 5.5 13.3962 5.5 12C5.5 10.6038 4.68248 9.39856 3.5 8.83682V6C3.5 5.72385 3.72385 5.5 4 5.5H20C20.2761 5.5 20.5 5.72385 20.5 6V8.83682C19.3175 9.39856 18.5 10.6038 18.5 12C18.5 13.3962 19.3175 14.6014 20.5 15.1632V18Z"
-                                  fill="#141C24"/>
-                            </svg>
+                        <TicketIcon />
                         <p>Билеты</p>
                     </span>
                 </div>
 
                 <span className={"desktop-navigation-span"}>
-                    <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-                        <path
-                            d="M14.4706 20.325C14.4706 19.9522 14.1684 19.65 13.7956 19.65H4.82353C4.54739 19.65 4.32353 19.4261 4.32353 19.15V4.85C4.32353 4.57386 4.54739 4.35 4.82353 4.35H13.7956C14.1684 4.35 14.4706 4.04779 14.4706 3.675C14.4706 3.30221 14.1684 3 13.7956 3H5C3.89543 3 3 3.89543 3 5V19C3 20.1046 3.89543 21 5 21H13.7956C14.1684 21 14.4706 20.6978 14.4706 20.325ZM12.8947 11.64C12.5468 11.64 12.2647 11.9221 12.2647 12.27C12.2647 12.6179 12.5468 12.9 12.8947 12.9H18.4412L16.2698 15.1148C16.0056 15.3843 16.0056 15.8157 16.2698 16.0852C16.5415 16.3624 16.9879 16.3624 17.2596 16.0852L20.3137 12.9701C20.6949 12.5812 20.6949 11.9588 20.3137 11.5699L17.2596 8.4548C16.9879 8.17764 16.5415 8.17764 16.2698 8.4548C16.0056 8.72431 16.0056 9.15569 16.2698 9.4252L18.4412 11.64H12.8947Z"
-                            fill="#141C24"/>
-                        </svg>
-
+                    <ExitIcon />
                     <p>Выйти</p>
                 </span>
                 <span className={"desktop-navigation-span"}>
-                    <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-                        <path
-                            d="M12 21C9.48333 21 7.35417 20.1292 5.6125 18.3875C3.87083 16.6458 3 14.5167 3 12C3 9.69999 3.75 7.70415 5.25 6.01249C6.75 4.32082 8.66667 3.33332 11 3.04998C11.2167 3.01665 11.4083 3.04582 11.575 3.13748C11.7417 3.22915 11.875 3.34998 11.975 3.49998C12.075 3.64998 12.1292 3.82498 12.1375 4.02498C12.1458 4.22499 12.0833 4.41665 11.95 4.59999C11.6667 5.03332 11.4542 5.49165 11.3125 5.97498C11.1708 6.45832 11.1 6.96665 11.1 7.49999C11.1 8.99999 11.625 10.275 12.675 11.325C13.725 12.375 15 12.9 16.5 12.9C17.0167 12.9 17.5292 12.825 18.0375 12.675C18.5458 12.525 19 12.3167 19.4 12.05C19.5833 11.9333 19.7708 11.8792 19.9625 11.8875C20.1542 11.8958 20.325 11.9417 20.475 12.025C20.6417 12.1083 20.7708 12.2333 20.8625 12.4C20.9542 12.5667 20.9833 12.7667 20.95 13C20.7167 15.3 19.7375 17.2083 18.0125 18.725C16.2875 20.2417 14.2833 21 12 21ZM12 19C13.4667 19 14.7833 18.5958 15.95 17.7875C17.1167 16.9792 17.9667 15.925 18.5 14.625C18.1667 14.7083 17.8333 14.775 17.5 14.825C17.1667 14.875 16.8333 14.9 16.5 14.9C14.45 14.9 12.7042 14.1792 11.2625 12.7375C9.82083 11.2958 9.1 9.54999 9.1 7.49999C9.1 7.16665 9.125 6.83332 9.175 6.49999C9.225 6.16665 9.29167 5.83332 9.375 5.49998C8.075 6.03332 7.02083 6.88332 6.2125 8.04999C5.40417 9.21665 5 10.5333 5 12C5 13.9333 5.68333 15.5833 7.05 16.95C8.41667 18.3167 10.0667 19 12 19Z"
-                            fill="#97A1AF"/>
-                        </svg>
+                    <SwitchMode />
                 </span>
             </div>
         </header>
