+++ ShiftCinema/src/styles/class-card-style.css
@@ -18,11 +18,13 @@
 .image-container  {
     position: relative;
     width: 100%;
+    height: 300px;
 }
 
 .image-container img {
     width: 100%;
-    height: auto;
+    height: 100%;
+    object-fit: cover;
     display: block;
     border-radius: 12px;
 }
@@ -37,18 +39,21 @@
     flex-direction: column;
     align-items: flex-end;
     box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+    background-color: rgba(128, 128, 128, 1);
+    color: white;
 }
 
 .image-container div p {
     font-size: 16px;
     font-weight: bold;
     white-space: nowrap;
+    margin: 0;
 }
 
 .film-title {
     font-size: 24px;
     font-weight: bold;
-    margin: 15px 20px 5px 20px;
+    margin: 15px 20px 0 20px;
 }
 
 .film-type {
