+++ ShiftCinema/src/styles/header-style.css
@@ -14,6 +14,9 @@
     flex-direction: column;
     align-items: center;
     gap: 10px;
+    z-index: 1000;
+    background: rgba(36, 36, 36, 255);
+    border-bottom: 1px solid rgba(255, 255, 255, 0.3);
 }
 
 .header-container{
